qa-chatbot/requirements.txt
google-generativeai
chainlit
python-dotenv
os
